Poll the absolut path of the manifest.json file
Instead of using a relative path to manifest.json use an absolut path.
This makes sure that we detect that Home Assistant Core is available
also on our Apps, which navigate to `/auth/authorize?...` by default.
The relative path then polls the wrong location `/auth/manifest.json`.

With the absolut path we make sure that the correct location is polled
on our Apps as well.