Add proxy for network and resolution endpoints (#124)
* Add proxy for network and resolution endpoints

Forward the requests to the network and resolution endpoints to the
Supervisor. This will allow the landing page to get insights when
the Supervisor detects issues on first startup (e.g. problematic
DNS servers). The network endpoint will allow to resolve these issues.

Note: This makes the two mentioned endpoints available unauthenticated.
However, the landing page will be replaced with Core as soon as it has
been downloaded, so this unauthenticated forwarding will only be during
a brief phase on initial setup. A bad actor could also just setup a new
user from the onboarding page at this point in time. So this effectly
does not change the security posture.

* Be less verbose in proxying