Check host network connectivity (#128)
* Check host network connectivity

Check host network connectivity and offer DNS alternatives in case host
internet is not available. This allows users with broken DNS setup
to configure the host network with a working DNS.

Requires https://github.com/home-assistant/supervisor/pull/5321.

* Apply suggestions from code review

Co-authored-by: Bram Kragten <mail@bramkragten.nl>

* Remove network issue warning once resolved

* Make const var again

All consts need to be assigned on creation. This is not possible in this
case.

* Improve wording

* Replace try with use

* Improve network issue wording

* Rename function, call schedule outside of then()

The idea is to call schedule() always, no matter if there is an error
or not. So call it outside of the then() block.

* Use let

---------

Co-authored-by: Bram Kragten <mail@bramkragten.nl>