Revert frontend version stamp to force update by workflow (#152)
The release assets merged in #149 are wrong because the workflow fix
from #150 was merged after the PR was created, so the workflow assumed
the assets are the latest and don't have to be updated. Lower the
frontend version to force the workflow to update the assets to latest
from the correct artifacts.